+++ packages/orm/src/test-utils/test-models.ts
@@ -5,6 +5,7 @@ import { modelToZod } from 'orchid-orm-schema-to-zod';
 export const Model = createModel({
   columnTypes: {
     ...columnTypes,
+    text: (min = 0, max = Infinity) => columnTypes.text(min, max),
     timestamp() {
       return columnTypes.timestamp().parse((input) => new Date(input));
     },