+++ packages/pqb/src/test-utils/test-utils.ts
@@ -24,6 +24,7 @@ export const db = createDb({
   adapter,
   columnTypes: {
     ...columnTypes,
+    text: (min = 0, max = Infinity) => columnTypes.text(min, max),
     timestamp() {
       return columnTypes.timestamp().parse((input) => new Date(input));
     },