+++ packages/schema-to-zod/CHANGELOG.md
@@ -1,5 +1,12 @@
 # orchid-orm-schema-to-zod
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - pqb@0.7.3
+
 ## 0.1.2
 
 ### Patch Changes